GLK: Merge Fonts class into Screen class
