+++ engines/glk/frotz/screen.h
