+++ engines/glk/screen.h
@@ -23,24 +23,56 @@
 #ifndef GLK_DRAW_H
 #define GLK_DRAW_H
 
+#include "common/archive.h"
+#include "common/array.h"
 #include "graphics/screen.h"
-#include "glk/fonts.h"
+#include "graphics/font.h"
+#include "glk/utils.h"
 
 namespace Glk {
+
+#define FONTS_TOTAL 8
 
 enum CaretShape {
 	SMALL_DOT = 0, FAT_DOT = 1, THIN_LINE = 2, FAT_LINE = 3, BLOCK = 4
 };
 
+enum FACES { MONOR, MONOB, MONOI, MONOZ, PROPR, PROPB, PROPI, PROPZ };
+enum TYPES { MONOF, PROPF };
+enum STYLES { FONTR, FONTB, FONTI, FONTZ };
+
 /**
  * Screen surface class
  */
-class Screen : public Graphics::Screen, public Fonts {
+class Screen : public Graphics::Screen {
+private:
+	Common::Array<const Graphics::Font *> _fonts;
+private:
+	/**
+	 * Load all the fonts
+	 */
+	bool loadFonts();
+
+	/**
+	 * Load a single font
+	 */
+	const Graphics::Font *loadFont(FACES face, Common::Archive *archive,
+		double size, double aspect, int style);
+public:
+	/**
+	 * Return the font Id for a given name
+	 */
+	static FACES getFontId(const Common::String &name);
 public:
 	/**
 	 * Constructor
 	 */
-	Screen() : Graphics::Screen(), Fonts(this) {}
+	Screen();
+
+	/**
+	 * Destructor
+	 */
+	virtual ~Screen();
 
 	/**
 	 * Fills the screen with a given rgb color
@@ -58,6 +90,44 @@ public:
 	 *      and the X position is in multiples of GLI_SUBPIX
 	 */
 	void drawCaret(const Point &pos);
+
+	/**
+	 * Draws a string using the specified font at the given co-ordinates
+	 * @param pos       Position for the bottom-left corner the text will be drawn with
+	 * @param fontIdx   Which font to use
+	 * @param rgb       RGB tuplet specifying the text color
+	 * @param text      The text to draw
+	 * @param spw       ??
+	 */
+	int drawString(const Point &pos, int fontIdx, const byte *rgb, const Common::String &text, int spw = 0);
+
+	/**
+	 * Draws a unicode string using the specified font at the given co-ordinates
+	 * @param pos       Position for the bottom-left corner the text will be drawn with
+	 * @param fontIdx   Which font to use
+	 * @param rgb       RGB tuplet specifying the text color
+	 * @param text      The text to draw
+	 * @param spw       ??
+	 */
+	int drawStringUni(const Point &pos, int fontIdx, const byte *rgb, const Common::U32String &text, int spw = 0);
+
+	/**
+	 * Get the width in pixels of a string
+	 * @param fontIdx   Which font to use
+	 * @param text      Text to get the width of
+	 * @param spw       Delta X
+	 * @returns         Width of string multiplied by GLI_SUBPIX
+	 */
+	size_t stringWidth(int fontIdx, const Common::String &text, int spw = 0);
+
+	/**
+	 * Get the width in pixels of a unicode string
+	 * @param fontIdx   Which font to use
+	 * @param text      Text to get the width of
+	 * @param spw       Delta X
+	 * @returns         Width of string multiplied by GLI_SUBPIX
+	 */
+	size_t stringWidthUni(int fontIdx, const Common::U32String &text, int spw = 0);
 };
 
 } // End of namespace Glk
